+++ src/lib/components/ui/theme-toggle/index.ts
@@ -1,0 +1,5 @@
+import ThemeToggle from './theme-toggle.svelte';
+
+export {
+	ThemeToggle
+};